refactor(core): deep architectural split for improved modularity
- Extracted generic utilities to utils/utils.js.
- Extracted popup management to utils/popup-manager.js.
- Extracted Bridge UI logic to scripts/bridge-ui.js.
- Extracted track processing logic to scripts/track-processor.js.
- Extracted YT Music API parsing to scripts/yt-music-parser.js.
- Refactored bridge.js, content.js, yt-music-api.js, and ui-helper.js to use new modules.
- Updated and expanded unit test suite.
